requirements.txt
apache-airflow[postgres]==3.0.1
apache-airflow-providers-common-sql==1.27.1
apache-airflow-providers-postgres==6.2.0
asyncpg
ccxt
psycopg2-binary==2.9.10
polars
python-dotenv